+++ Modules/Intent.RelationalDbSchemaImporter.Contracts/Enums/AttributeNameConvention.cs
@@ -2,6 +2,6 @@ namespace Intent.RelationalDbSchemaImporter.Contracts.Enums;
 
 public enum AttributeNameConvention
 {
-    LanguageCompliant,
-    PreserveOriginal
+    Default,
+    ColumnName
 }